+++ cpp/bisMemoryManagement.h
@@ -66,7 +66,7 @@ namespace bisMemoryManagement {
    * @param operation description of the operation
    * @param owner owner object for this memory
    */
-  unsigned char* allocate_memory(int sz,std::string name="",std::string operation="",bisObject* owner=0);
+  unsigned char* allocate_memory(long sz,std::string name="",std::string operation="",bisObject* owner=0);
 
   /**
    * Release memory at location specified by pointer
@@ -100,7 +100,7 @@ namespace bisMemoryManagement {
    * @param input location of input data
    * @param length number of bytes to copy
    */
-  void copy_memory(unsigned char* output,unsigned char* input,int length);
+  void copy_memory(unsigned char* output,unsigned char* input,long length);
